requirements.txt
langgraph>=0.3.25
langchain>=0.3.14
langchain-core>=0.3.30
langchain-groq>=0.2.0
langchain-community>=0.3.0
python-dotenv>=1.0.0
tavily-python>=0.5.0
